a,b=map(float,input().split()) ave=(a+b)/2 print("%.4f"%ave)

+++ Find_the_average_of_two_numbers.py
@@ -1,3 +1,3 @@
 a,b=map(float,input().split())
-ave=(float)(a+b)/2
+ave=(a+b)/2
 print("%.4f"%ave)
